refactor: list issue
